Moved token policy handling to seperate files.
