+++ desec.go
@@ -55,10 +55,11 @@ type Client struct {
 	common service // Reuse a single struct instead of allocating one for each service on the heap.
 
 	// Services used for talking to different parts of the deSEC API.
-	Account *AccountService
-	Tokens  *TokensService
-	Records *RecordsService
-	Domains *DomainsService
+	Account       *AccountService
+	Tokens        *TokensService
+	Records       *RecordsService
+	Domains       *DomainsService
+	TokenPolicies *TokenPoliciesService
 }
 
 // New creates a new Client.
@@ -81,6 +82,7 @@ func New(token string, opts ClientOptions) *Client {
 	client.Tokens = (*TokensService)(&client.common)
 	client.Records = (*RecordsService)(&client.common)
 	client.Domains = (*DomainsService)(&client.common)
+	client.TokenPolicies = (*TokenPoliciesService)(&client.common)
 
 	return client
 }
